+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.30.0 — unreleased
+## 0.30.0 — 2026-09-15
 
 ### Fixed